requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
yt-dlp==2024.3.10
pymongo==4.6.1
motor==3.3.2
python-multipart==0.0.9 